Move combine step into endgame post processing
Other diseases will need to use this step
